Fix case of "PolyHook" in cmake config file. (#64)
This ensures that find_library(PolyHook_2 CONFIG REQUIRE) functions
properly on systems with case sensitive paths.

In addition, also uses a path to the target file that is relative to
the config file, which is helpful in case the cmake files are moved to
a different location (e.g. for vcpkg).