Add cmake compile options
POLYHOOK_COMPILE_32 = add C and C++ flags to compile for 32 bits
POLYHOOK_DISABLE_IOSTREAM = Replace iostream with alternative. It was crashing on my linux build for some reason.
POLYHOOK_DISABLE_INSTRUCTION_PRINTING = Disable the printing of instructions in the debug log. Could make a difference when hooking hundreds of functions.

Add a bunch of ifdef in IHook.hpp to prevent linux compile errors.

Add an x86Detour constructor that accepts 32bit ints and upcasts them to 64bit behind the scenes.